Support for custom templates.
Added support for customizable templates location.
Another property can be used to specify the location of the custom text temlates for class generation.
Currently, this is supported in the comman-line interface as 4-th argument.
The old functionality is preserved, if not supplied, the generator will read from the standard resources directory.